pyproject.toml
[project]
name = "xScanner"
version = "0.1.3"
description = "OCR and LLM-based bullion bar data extraction"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "aXedras"}
authors = [
    {name = "aXedras", email = "marco.persi@axedras.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: aXedras",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core dependencies - always needed
    "requests==2.32.5",
    "Pillow>=10.3.0",  # Image processing (fixed CVE-2023-50447, CVE-2024-28219)
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
# Cloud-based OCR strategies (lightweight)
cloud = [
    "openai==1.35.0",  # ChatGPT Vision
    "google-generativeai>=0.3.0",  # Gemini Flash
]

# Local OCR strategies (heavy dependencies)
local = [
    "paddleocr>=3.3.0",  # Updated for opencv-python compatibility
    "paddlepaddle>=2.6.2",
    "opencv-python>=4.8.1.78",  # Fixed CVE-2023-4863 (libwebp)
    "numpy<2.0",
]

# All OCR strategies (for development/testing)
all-strategies = [
    "xScanner[cloud,local]",
]

server = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.5.0",
    "redis>=5.0.0",
    "aiofiles>=23.2.0",
    "httpx>=0.25.0",
    "pypdf>=4.0.0",
    "PyMuPDF>=1.24.0",
]

dev = [
    "ruff>=0.1.8",
    "mypy>=1.7.0",
    "types-requests>=2.31.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.25.0",
    "pre-commit>=3.6.0",
]

[tool.ruff]
target-version = "py310"
line-length = 100
extend-exclude = [
    "tmp",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # function call in default arg (FastAPI dependency injection)
]

[tool.ruff.lint.isort]
known-first-party = ["xscanner"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.10"
warn_return_any = false
warn_unused_ignores = true
disallow_untyped_defs = false
ignore_missing_imports = true
check_untyped_defs = false

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["xscanner*"]
exclude = ["tests*", "build*", "dist*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short -m 'not e2e'"
pythonpath = ["src"]
filterwarnings = [
    # PyMuPDF/PyMuPDFb (SWIG) currently emits these on Python 3.12+
    "ignore:builtin type SwigPyPacked has no __module__ attribute:DeprecationWarning",
    "ignore:builtin type SwigPyObject has no __module__ attribute:DeprecationWarning",
    "ignore:builtin type swigvarlink has no __module__ attribute:DeprecationWarning",
]
markers = [
    "integration: integration tests (no external API keys required)",
    "e2e: end-to-end tests (may require external services and/or API keys)",
]
